requirements.txt
python-binance
python-telegram-bot
numpy
matplotlib
requests